Added MessageFiltering strategy ( so the msgQueue is filled - and read by the state - only with messages that it should accept ).
Bug?: On "closing" the listening thread still waits one more time before closing ?
Warning: Missing filling of the filtering method on ALL STATEs or it wont work.
NOTA: to test with methods onside each State instead of lambda (so code is cleaner if "big filtering functions")
NOTA2: The filtering shall be a lookup inside an unordered_set (if count > 0 ). WIP and to update with command code later !!